examples: rename 09_wave_visualization_Moon -> 11_wave_visualization_Moon and update
