+++ examples/11_wave_visualization_Moon/sim/ISSI_MOON_M1_atten.bm
@@ -1,0 +1,187 @@
+# Velocity model for the moon according to Garcia et al., 2019. “Lunar 
+# Seismology: An Update on Interior Structure Models.” 
+# 
+# New values for attenuation based on values from the literature. 
+#
+NAME        ISSI_M1_atten
+ANELASTIC   true
+ANISOTROPIC false
+UNITS       km
+COLUMNS     depth vp vs rho qmu qkappa
+    0.00 4.04 2.32 3.5 5000 1000000
+    10.00 4.65 2.68 3.5 5000 1000000
+    20.00 5.26 3.05 3.5 5000 1000000
+    30.00 5.90 3.42 3.5 5000 1000000
+    40.00 6.51 3.79 3.5 5000 1000000
+    50.00 7.10 4.12 3.5 5000 1000000
+    50.00 7.10 4.12 3.5 2133 1000000
+    60.00 7.57 4.38 3.5 2133 1000000
+    70.00 7.62 4.42 3.5 2133 1000000
+    80.00 7.63 4.42 3.5 2133 1000000
+    90.00 7.64 4.42 3.5 2133 1000000
+    100.00 7.64 4.42 3.5 2133 1000000
+    110.00 7.64 4.41 3.5 2133 1000000
+    120.00 7.64 4.41 3.5 2133 1000000
+    130.00 7.64 4.41 3.5 2133 1000000
+    140.00 7.64 4.40 3.5 2133 1000000
+    150.00 7.63 4.40 3.5 2133 1000000
+    160.00 7.63 4.39 3.5 2133 1000000
+    170.00 7.63 4.39 3.5 2133 1000000
+    180.00 7.63 4.38 3.5 2133 1000000
+    190.00 7.62 4.38 3.5 2133 1000000
+    200.00 7.62 4.38 3.5 2133 1000000
+    210.00 7.62 4.38 3.5 2133 1000000
+    220.00 7.62 4.38 3.5 2133 1000000
+    230.00 7.62 4.37 3.5 2133 1000000
+    240.00 7.62 4.37 3.5 2133 1000000
+    250.00 7.62 4.37 3.5 2133 1000000
+    260.00 7.62 4.37 3.5 2133 1000000
+    270.00 7.62 4.37 3.5 2133 1000000
+    280.00 7.63 4.37 3.5 2133 1000000
+    290.00 7.64 4.37 3.5 2133 1000000
+    300.00 7.64 4.37 3.5 2133 1000000
+    310.00 7.65 4.37 3.5 2133 1000000
+    320.00 7.65 4.37 3.5 2133 1000000
+    330.00 7.66 4.38 3.5 2133 1000000
+    340.00 7.67 4.38 3.5 2133 1000000
+    350.00 7.68 4.38 3.5 2133 1000000
+    360.00 7.69 4.39 3.5 2133 1000000
+    370.00 7.70 4.39 3.5 2133 1000000
+    380.00 7.71 4.39 3.5 2133 1000000
+    390.00 7.72 4.40 3.5 2133 1000000
+    400.00 7.73 4.40 3.5 2133 1000000
+    410.00 7.74 4.41 3.5 419 1000000
+    420.00 7.75 4.41 3.5 419 1000000
+    430.00 7.76 4.42 3.5 419 1000000
+    440.00 7.77 4.42 3.5 419 1000000
+    450.00 7.78 4.43 3.5 419 1000000
+    460.00 7.79 4.43 3.5 419 1000000
+    470.00 7.80 4.44 3.5 419 1000000
+    480.00 7.81 4.44 3.5 419 1000000
+    490.00 7.82 4.45 3.5 419 1000000
+    500.00 7.84 4.45 3.5 419 1000000
+    510.00 7.85 4.46 3.5 419 1000000
+    520.00 7.86 4.46 3.5 419 1000000
+    530.00 7.87 4.47 3.5 419 1000000
+    540.00 7.88 4.47 3.5 419 1000000
+    550.00 7.89 4.48 3.5 419 1000000
+    560.00 7.90 4.48 3.5 419 1000000
+    570.00 7.91 4.49 3.5 419 1000000
+    580.00 7.92 4.49 3.5 419 1000000
+    590.00 7.93 4.50 3.5 419 1000000
+    600.00 7.94 4.51 3.5 419 1000000
+    610.00 7.96 4.51 3.5 419 1000000
+    620.00 7.97 4.52 3.5 419 1000000
+    630.00 7.98 4.52 3.5 419 1000000
+    640.00 7.99 4.53 3.5 419 1000000
+    650.00 8.00 4.54 3.5 419 1000000
+    660.00 8.01 4.54 3.5 419 1000000
+    670.00 8.02 4.55 3.5 419 1000000
+    680.00 8.03 4.55 3.5 419 1000000
+    690.00 8.04 4.56 3.5 419 1000000
+    700.00 8.05 4.57 3.5 419 1000000
+    710.00 8.06 4.57 3.5 419 1000000
+    720.00 8.08 4.58 3.5 419 1000000
+    730.00 8.09 4.59 3.5 419 1000000
+    740.00 8.10 4.59 3.5 419 1000000
+    750.00 8.11 4.60 3.5 419 1000000
+    760.00 8.13 4.61 3.5 419 1000000
+    770.00 8.14 4.62 3.5 419 1000000
+    780.00 8.15 4.62 3.5 419 1000000
+    790.00 8.16 4.63 3.5 419 1000000
+    800.00 8.17 4.64 3.5 419 1000000
+    810.00 8.21 4.67 3.5 121 1000000
+    820.00 8.22 4.67 3.5 121 1000000
+    830.00 8.24 4.68 3.5 121 1000000
+    840.00 8.24 4.69 3.5 121 1000000
+    850.00 8.24 4.69 3.5 121 1000000
+    860.00 8.24 4.69 3.5 121 1000000
+    870.00 8.24 4.69 3.5 121 1000000
+    880.00 8.24 4.69 3.5 121 1000000
+    890.00 8.23 4.69 3.5 121 1000000
+    900.00 8.23 4.68 3.5 121 1000000
+    910.00 8.21 4.66 3.5 121 1000000
+    920.00 8.17 4.63 3.5 121 1000000
+    930.00 8.11 4.60 3.5 121 1000000
+    940.00 8.05 4.57 3.5 121 1000000
+    950.00 7.99 4.54 3.5 121 1000000
+    960.00 7.92 4.51 3.5 121 1000000
+    970.00 7.84 4.47 3.5 121 1000000
+    980.00 7.76 4.44 3.5 121 1000000
+    990.00 7.69 4.40 3.5 121 1000000
+    1000.00 7.66 4.37 3.5 121 1000000
+    1010.00 7.63 4.36 3.5 121 1000000
+    1020.00 7.60 4.36 3.5 121 1000000
+    1030.00 7.60 4.38 3.5 121 1000000
+    1040.00 7.60 4.39 3.5 121 1000000
+    1050.00 7.64 4.40 3.5 121 1000000
+    1060.00 7.65 4.40 3.5 121 1000000
+    1070.00 7.62 4.38 3.5 121 1000000
+    1080.00 7.59 4.37 3.5 121 1000000
+    1090.00 7.57 4.37 3.5 121 1000000
+    1100.00 7.53 4.36 3.5 121 1000000
+    1110.00 7.52 4.35 3.5 121 1000000
+    1120.00 7.49 4.37 3.5 121 1000000
+    1130.00 7.49 4.39 3.5 121 1000000
+    1140.00 7.50 4.43 3.5 121 1000000
+    1150.00 7.51 4.45 3.5 121 1000000
+    1160.00 7.55 4.47 3.5 121 1000000
+    1170.00 7.56 4.49 3.5 121 1000000
+    1180.00 7.52 4.46 3.5 121 1000000
+    1190.00 7.54 4.45 3.5 121 1000000
+    1200.00 7.55 4.45 3.5 121 1000000
+    1210.00 7.36 4.36 3.5 121 1000000
+    1220.00 7.29 4.31 3.5 121 1000000
+    1230.00 7.24 4.24 3.5 121 1000000
+    1240.00 7.20 4.19 3.5 121 1000000
+    1250.00 7.12 4.12 3.5 121 1000000
+    1260.00 7.05 4.00 3.5 121 1000000
+    1270.00 6.99 3.93 3.5 121 1000000
+    1280.00 6.94 3.84 3.5 121 1000000
+    1290.00 6.88 3.73 3.5 121 1000000
+    1300.00 6.80 3.57 3.5 121 1000000
+    1310.00 6.71 3.40 3.5 121 1000000
+    1320.00 6.62 3.21 3.5 121 1000000
+    1330.00 6.56 2.62 3.5 121 1000000
+    1340.00 6.56 2.62 3.5 121 1000000
+    1340.00 6.50 0.00 3.5 0 1000000
+    1350.00 6.36 0.00 3.5 0 1000000
+    1360.00 6.22 0.00 3.5 0 1000000
+    1370.00 6.12 0.00 3.5 0 1000000
+    1380.00 5.93 0.00 3.5 0 1000000
+    1390.00 5.77 0.00 3.5 0 1000000
+    1400.00 5.32 0.00 3.5 0 1000000
+    1410.00 5.32 0.00 3.5 0 1000000
+    1420.00 5.32 0.00 3.5 0 1000000
+    1430.00 5.32 0.00 3.5 0 1000000
+    1440.00 5.32 0.00 3.5 0 1000000
+    1450.00 5.32 0.00 3.5 0 1000000
+    1460.00 5.32 0.00 3.5 0 1000000
+    1470.00 5.32 0.00 3.5 0 1000000
+    1480.00 5.32 0.00 3.5 0 1000000
+    1490.00 5.32 0.00 3.5 0 1000000
+    1500.00 5.32 0.00 3.5 0 1000000
+    1510.00 5.32 0.00 3.5 0 1000000
+    1520.00 5.32 0.00 3.5 0 1000000
+    1530.00 5.32 0.00 3.5 0 1000000
+    1540.00 5.32 0.00 3.5 0 1000000
+    1550.00 5.32 0.00 3.5 0 1000000
+    1560.00 5.32 0.00 3.5 0 1000000
+    1570.00 5.32 0.00 3.5 0 1000000
+    1580.00 5.32 0.00 3.5 0 1000000
+    1590.00 5.32 0.00 3.5 0 1000000
+    1600.00 5.32 0.00 3.5 0 1000000
+    1610.00 5.32 0.00 3.5 0 1000000
+    1620.00 5.32 0.00 3.5 0 1000000
+    1630.00 5.32 0.00 3.5 0 1000000
+    1640.00 5.32 0.00 3.5 0 1000000
+    1650.00 5.32 0.00 3.5 0 1000000
+    1660.00 5.32 0.00 3.5 0 1000000
+    1670.00 5.32 0.00 3.5 0 1000000
+    1680.00 5.32 0.00 3.5 0 1000000
+    1690.00 5.32 0.00 3.5 0 1000000
+    1700.00 5.32 0.00 3.5 0 1000000
+    1710.00 5.32 0.00 3.5 0 1000000
+    1720.00 5.32 0.00 3.5 0 1000000
+    1730.00 5.32 0.00 3.5 0 1000000
+    1737.10 5.32 0.00 3.5 0 1000000
